requirements.txt
absl-py==1.3.0
ale-py==0.8.0
AutoROM==0.4.2
AutoROM.accept-rom-license==0.4.2
brotlipy==0.7.0
certifi==2022.9.24
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
cloudpickle==2.2.0
contourpy==1.0.6
cryptography @ file:///croot/cryptography_1665612644927/work
cycler==0.11.0
Cython==0.29.34
dm-env==1.6
dm-tree==0.1.8
Farama-Notifications==0.0.4
fasteners==0.18
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.38.0
glfw==2.5.5
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.28.1
h5py==3.8.0
idna @ file:///croot/idna_1666125576474/work
imageio==2.28.1
importlib-metadata==4.13.0
importlib-resources==5.10.1
jax-jumpy==1.0.0
kiwisolver==1.4.4
labmaze==1.0.6
lxml==4.9.2
matplotlib==3.6.2
mjrl @ git+https://github.com/aravindr93/mjrl@3871d93763d3b49c4741e6daeaebbc605fe140dc
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
mujoco==2.3.3
numpy @ file:///croot/numpy_and_numpy_base_1668593735768/work
opencv-python==4.6.0
packaging==21.3
pandas==1.5.2
PettingZoo==1.22.3
Pillow==9.2.0
protobuf==4.22.3
pybullet==3.2.5
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
PyOpenGL==3.1.6
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
sb3-contrib==1.7.0
scipy==1.10.1
seaborn==0.12.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
stable-baselines3 @ git+https://github.com/carlosluis/stable-baselines3@3f75a8aedc9ee183238919d7a71734a0ef68872c
termcolor==2.3.0
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tqdm==4.64.1
typing_extensions @ file:///croot/typing_extensions_1669924550328/work
urllib3 @ file:///croot/urllib3_1666298941550/work
zipp==3.11.0
